Types and Functions

Types

IDFCurves.GeneralScaling — Type

GeneralScaling(d₀::Real, μ₀::Real, σ₀::Real, ξ::Real, α::Real, δ::Real)

Construct a GeneralScaling distribution type.

Details

TODO

$\mu_d = \mu_0 \left(\frac{d+\delta}{d_0+\delta} \right) ^{-\alpha} \qquad \sigma_d = \sigma_0 \left(\frac{d+\delta}{d_0+\delta} \right) ^{-\alpha} \qquad \xi_d = \xi$

References

Koutsoyiannis, D., Kozonis, D. and Manetas, A. (1998). A mathematical framework for studying rainfall intensity-duration-frequency relationships, Journal of Hydrology, 206(1-2), 118-135, https://doi.org/10.1016/S0022-1694(98)00097-3.

IDFCurves.IDFdata — Method

IDFdata(df::DataFrame, year_id::String, duration::Dict{String, T} where T<:Real)

Construct a IDFdata structure from a DataFrame.

Details

year_id: string indicating the year column

duration: Dictionary mapping the dataframe id to duration.

See the tutorial for an example.

IDFCurves.SimpleScaling — Type

SimpleScaling(d₀::Real, μ₀::Real, σ₀::Real, ξ::Real, α::Real)

Construct a simple scaling distribution type.

Details

TODO

$\mu_d = \mu_0 \left(\frac{d}{d_0} \right) ^{-\alpha} \qquad \sigma_d = \sigma_0 \left(\frac{d}{d_0} \right) ^{-\alpha} \qquad \xi_d = \xi$

References

Koutsoyiannis, D., Kozonis, D. and Manetas, A. (1998). A mathematical framework for studying rainfall intensity-duration-frequency relationships, Journal of Hydrology, 206(1-2), 118-135, https://doi.org/10.1016/S0022-1694(98)00097-3.

IDFCurves.UniversalScaling — Type

UniversalScaling(d₀::Real, μ₀::Real, σ₀::Real, ξ::Real, α::Real, δ::Real, τ::Real)

Construct a UniversalScaling distribution type.

Details

TODO

$\mu_d = \mu_0 \left(\frac{(d+\delta)^{-\alpha}+\tau}{(d_0+\delta)^{-\alpha}+\tau} \right) \qquad \sigma_d = \sigma_0 \left(\frac{(d+\delta)^{-\alpha}+\tau}{(d_0+\delta)^{-\alpha}+\tau} \right) \qquad \xi_d = \xi$

Functions

IDFCurves.dataset — Method

dataset(name::String)::DataFrame

Load the dataset associated with name.

Details

Some datasets are available using the following names:

1108446: short duration rainfall intensity-duration-frequency data recorded at Vancouver Harbour CS

6158731: short duration rainfall intensity-duration-frequency data recorded at the Toronto Leaster B. Pearson Internation Airport

702S006: short duration rainfall intensity-duration-frequency data recorded at the Montreal Pierre-Elliott-Trudeau Internation Airport

These datasets have been retrieved from the Environment and Climate Change Canada website.

Examples

julia> IDFCurves.dataset("702S006")

Base.show — Method

Base.show(io::IO, obj::IDFdata)

Override of the show function for the objects of type IDFdata.

IDFCurves._common_years — Method

_common_years(data::IDFdata, tags::AbstractVector{<:AbstractString})

Extract the common years vector of the records in data corresponding to the tags.

IDFCurves._restrict_years — Method

_restrict_years(data::IDFdata, years::AbstractVector{<:Integer}; tags = gettag(data))

Return a new IDFdata struct restrained to the years for each tags.

IDFCurves._validation_tag — Method

_validation_tag(data::IDFdata, tag_out)

The function returns the tag if it exists in data, otherwise, the function throws an error.

IDFCurves._validation_tag — Method

_validation_tag(data::IDFdata)

The tag corresponding to the minimal duration in `data` is returned.

IDFCurves.excludeduration — Method

excludeduration(data::IDFdata, tag_out::String)

Remove data of data corresponding to the duration identified by tag_out.

IDFCurves.getKendalldata — Method

getKendalldata(obj::IDFdata)

Computes the Kendall tau for each pair of durations for which obj contains data, and returns them in a DataFrame.

IDFCurves.getdata — Method

getdata(s::IDFdata, tag::String, year::Int)

Return the data corresponding to the duration tag tag and the year year.

IDFCurves.getdata — Method

getdata(s::IDFdata, tag::String)

Return the data vector corresponding to the duration tag tag.

IDFCurves.getdata — Method

getdata(s::IDFdata)

Return a dictionary containing the data vector for each duration tag.

IDFCurves.getduration — Method

getduration(s::IDFdata, tag::String)

Return the vector of durations for the duration tag tag.

IDFCurves.getduration — Method

getduration(s::IDFdata)

Return a dictionary containing the duration vector for each duration tag.

IDFCurves.gettag — Method

gettag(data::IDFdata, d::Real)

Return the tag corresponding to the duration d if it exists; throw an error otherwise.

IDFCurves.gettag — Method

gettag(data::IDFdata)

Return the tag list.

IDFCurves.getyear — Method

getyear(s::IDFdata, tag::String)

Return the vector of years for the duration tag tag.

IDFCurves.getyear — Method

getyear(s::IDFdata)

Return a dictionary containing the year vector for each duration tag.

IDFCurves.get_durations_labels — Method

get_durations_labels(durations::Vector{<:Real})

Returns a vector of labels (String) based on a vector of durations. The returned vector contains the labels whiwh will appear on the x-axis wwhen plotting the IDF curve.

IDFCurves.get_layers_IDFCurves — Method

getlayersIDFCurves(model::DependentScalingModel, Tvalues::Vector{<:Real}, durationsrange::Vector{<:Real})

Returns the layers associated to the IDF curves based on the given mode, for each return period in T_values.

IDFCurves.get_layers_pointwise_estimations — Method

getlayerspointwiseestimations(model::DependentScalingModel, Tvalues::Vector{<:Real}, durations::Vector{<:Real})

Returns the layers associated to pointwise estimations of the return levels at the differents durations, for every return period in T_values.

IDFCurves.plotIDFCurves — Method

plotIDFCurves(model::DependentScalingModel, data::IDFdata; ...)

Plots the IDF curves based on the given model. Pointwise estimations are added (crosses) to illustrate the fitting. Durations and return periods may be chosen by the user but have default values. If showconfidenceintervals = true, 95% confidence intervals associated to the pointwise estimations are represented.

IDFCurves.plotIDFCurves — Method

plotIDFCurves(model::DependentScalingModel; ...)

Pots the IDF curves based on the given model. Durations and return periods may be chosen by the user but have default values.

IDFCurves.qqplot — Method

qqplot(pd::Distribution, y::Vector{Float64})

Quantile-Quantile plot from ExtendedExtremes.jl

IDFCurves.qqplot — Method

qqplot(pd::Distribution, y::Vector{Float64})

Quantile-Quantile plot from ExtendedExtremes.jl

IDFCurves.qqplotci — Function

qqplotci(fm::MarginalScalingModel, α::Real=.05, H::PDMat{<:Real})

Quantile-Quantile plot along with the confidence/credible interval of level 1-α.

Details

This function uses the Hessian matrix H provided in the argument.

IDFCurves.qqplotci — Function

qqplotci(fm::DependentScalingModel, α::Real=.05, H::PDMat{<:Real})

Quantile-Quantile plot along with the confidence/credible interval of level 1-α.

Details

This function uses the Hessian matrix H provided in the argument.

IDFCurves.qqplotci — Function

qqplotci(fm::DependentScalingModel, data::IDFdata, d::Real, α::Real=.05)

Quantile-Quantile plot for the estimated distribution for duration d, along with the confidence interval of level 1-α.

IDFCurves.qqplotci — Function

qqplotci(fm::MarginalScalingModel, α::Real=.05)

Quantile-Quantile plot along with the confidence/credible interval of level 1-α.